Game breaking and individual files for quest
This update will probably be the last update in a long time. Though I
loved developing this, I do not have time or interest to continue with
Feudalism. I will not completely abandon it, but this wont be a priority
any more. Also the fact that so few people seems to have interest in
what I have created sends a clear message to me. If you would like to
help me with this please send me a issue with the label "co-develop"

Anyway: Here are what's new in this update!

+Created a help command (you can only see the commands you can execute)
+Cut each quest into its own file
+Enchanted the command "saveItem"
+Renamed command "setTeleport" to "setLocation" and enchanted it
+You can now toggle between see and not see the amount of xp you get
with "pxp"
+You now have to say which player you clear in "fullyclear"
+Fixed the SCZ command
+Fixed pick up for op
+Minor tweaks
+Bug fixes